Fixed a bug, img of partition fixed it

+++ os/src/Converter.h
@@ -41,6 +41,6 @@ public:
 	static const uint16_t MFT_SIZE_B;
 
 	static const uint8_t RESERVED_MFT;
-	static const uint16_t END_MARKER;
+	static const uint32_t END_MARKER;
 };
